+++ src/index.ts
@@ -1,5 +1,5 @@
 import { callable } from './common/utils.ts'
-import MarkdownItClass from './markdownit.ts'
+import MarkdownIt from './markdownit.ts'
 
 /**
  * Default package export.
@@ -11,6 +11,18 @@ import MarkdownItClass from './markdownit.ts'
  *
  * @category Main
  */
-const MarkdownIt = callable(MarkdownItClass)
+const MarkdownItCallable = callable(MarkdownIt)
 
-export default MarkdownIt
+export default MarkdownItCallable
+
+export type { default as MarkdownIt, MarkdownItPreset } from './markdownit.ts'
+export type { Delimiter, Env, MarkdownItOptions } from './types.ts'
+export type { default as Token } from './token.ts'
+export type { default as Ruler } from './ruler.ts'
+export type { default as Renderer, RendererRule } from './renderer.ts'
+export type { default as ParserCore } from './parser_core.ts'
+export type { default as StateCore } from './rules_core/state_core.ts'
+export type { default as ParserBlock } from './parser_block.ts'
+export type { default as StateBlock } from './rules_block/state_block.ts'
+export type { default as ParserInline } from './parser_inline.ts'
+export type { default as StateInline } from './rules_inline/state_inline.ts'